Move moire-gap-data out of entries/ to prevent auto-indexing
The post-summarizer vite plugin auto-generates entries/index.ts from
all files in entries/. Moving the raw gap data to lib/data/ prevents
it from being registered as a post entry.
